Minor refactoring on settings page
will be used in team page
